refactor(Result) : Result클래스를 통한 exception 처리
