+++ src/main/java/IO/Console.java
@@ -1,6 +1,7 @@
 package IO;
 
 import Record.*;
+import Result.ErrorCode;
 
 import java.util.Iterator;
 import java.util.Map;
