+++ src/main/java/IO/Output.java
@@ -1,6 +1,7 @@
 package IO;
 
 import Record.Record;
+import Result.ErrorCode;
 
 import java.util.Map;
 
@@ -9,5 +10,4 @@ public interface Output {
     void output(String s);
     void allCalcRecord(Map<Long, Record> map);
     void errorPrint(ErrorCode errorCode);
-
 }
